Ignore terraform.tfvars
It's risky to store terraform.tfvars in the repo since it might eventually have
sensitive information. This commit provides an example if needed but doesn't
commit the actual file to the repoIgnore terraform.tfvars

It's risky to store terraform.tfvars in the repo since it might eventually have
sensitive information. This commit provides an example if needed but doesn't
commit the actual file to the repo.